Updated Script & Added Mechanism to Upload /w PHP
The script has been changed to remove numerous unnecessary print
statements.

The HTML file is the base code that makes it possible to upload the CSV
to the web server.

The PHP will rename the CSV to 'wdata.csv' and then write it to the
folder where import.py is stored.  It will then run import.py, which
will parse this newly uploaded CSV and update the database as
appropriate.  It will finally delete the CSV and report out success or
failure.